+++ src/main/java/com/siteshkumar/bms/Service/Impl/PostServiceImpl.java
@@ -2,6 +2,7 @@ package com.siteshkumar.bms.Service.Impl;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import com.siteshkumar.bms.DTO.PostRequest;
 import com.siteshkumar.bms.DTO.PostResponse;
@@ -11,22 +12,24 @@ import com.siteshkumar.bms.Entity.UserEntity;
 import com.siteshkumar.bms.Mapper.PostMapper;
 import com.siteshkumar.bms.Repository.PostRepository;
 import com.siteshkumar.bms.Repository.UserRepository;
+import com.siteshkumar.bms.Security.AuthUtils;
+import com.siteshkumar.bms.Security.CustomUserDetails;
 import com.siteshkumar.bms.Service.PostService;
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class PostServiceImpl implements PostService{
     
     private final PostRepository postRepository;
     private final UserRepository userRepository;
-
-    public PostServiceImpl(PostRepository postRepository, UserRepository userRepository){
-        this.postRepository = postRepository;
-        this.userRepository = userRepository;
-    }
+    private final AuthUtils authUtils;
 
     @Override
     public PostResponse createPost(PostRequest dto){
-        UserEntity user = userRepository.findById(dto.getUserId())
+        CustomUserDetails currentUser = authUtils.getCurrentLoggedInUser();
+
+        UserEntity user = userRepository.findById(currentUser.getId())
                         .orElseThrow(() -> new RuntimeException("User not found!!!"));
 
         PostEntity post = new PostEntity();
@@ -46,6 +49,15 @@ public class PostServiceImpl implements PostService{
         PostEntity post = postRepository.findById(postId)
                         .orElseThrow(() -> new RuntimeException(" Post not found!!! "));
 
+        CustomUserDetails currentUser = authUtils.getCurrentLoggedInUser();
+
+        // System.out.println("POST OWNER = " + post.getAuthor().getUserId());
+        // System.out.println("CURRENT USER = " + currentUser.getId());
+        // System.out.println("***********************************************");
+        
+        if(! post.getAuthor().getUserId().equals(currentUser.getId()))
+            throw new AccessDeniedException("You are not allowed to update this post");
+
         if(dto.getContent() != null && !dto.getContent().isBlank())
             post.setContent(dto.getContent());
 
@@ -62,6 +74,15 @@ public class PostServiceImpl implements PostService{
     public void deletePost(Long postId){
         PostEntity post = postRepository.findById(postId)
                         .orElseThrow(() -> new RuntimeException("Post not found!!!"));
+
+        CustomUserDetails currentUser = authUtils.getCurrentLoggedInUser();
+        boolean isOwner = post.getAuthor().getUserId().equals(currentUser.getId());
+        boolean isAdmin = currentUser.getAuthorities()
+                                    .stream()
+                                    .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isOwner && !isAdmin)
+            throw new AccessDeniedException("You are not allowed to delete this post");
 
         postRepository.delete(post);
     }
